login

+++ addProduct/App.js
@@ -1,20 +1,35 @@
 class Product {
-  constructor(name, price) {
-    this.name = name;
-    this.price = price;
-  }
+
+    static contador = 0;
+
+    constructor(mark, name, price) {
+        Product.contador ++;
+        this.mark = mark
+        this.name = name;
+        this.price = price;
+        this.id = Product.contador;
+    }
+ 
 }
 
 class UI {
   addProduct(product) {
-      $("#product-list").append(`<div class="card text-center mb-4">
-                                    <div class="card-body">
-                                        <strong>Name</strong>: ${product.name}
-                                        <strong>Price</strong>: ${product.price}
-                                        <strong>Year</strong>: ${product.year}
-                                        <a href="#" class="btn btn-danger" name="delete">Delete</a>
-                                    </div>
-                                </div>`);
+      $("tbody").append(`<tr id="${product.id}">
+                            <td>${product.id}</td>
+                            <td>${product.mark}</td>
+                            <td>${product.name}</td>
+                            <td>${product.price}</td>
+                            <td>
+                                <button name="edit" type="button" class="btn btn-success" data-toggle="modal" data-target="#exampleModal">
+                                    <i class="fas fa-edit"></i>
+                                </button>
+                            </td>
+                            <td>
+                                <button name="delete" type="button" class="btn btn-danger" data-toggle="modal" data-target="#exampleModal">
+                                    <i class="fas fa-trash-alt"></i>
+                                </button>
+                            </td>
+                          </tr>`);
   }
 
   resetForm()
@@ -22,24 +37,52 @@ class UI {
         $("#product-form")[0].reset();
   }
 
+  resetModal()
+  {
+        $('.modal-title').empty();
+        $('.modal-body').empty();
+  }
+
   deleteProduct(e)
   {
-        if(e.name === "delete")
-        {
-            console.log($(e).parent().parent().remove());
+        this.resetModal();
+        $('.modal-title').append('Delete product');
+        $('.modal-body').append('are you sure you want to delete the product?');
+        $("#saveChanges").click(()=>{
+            $(e).parent().parent().remove();
             this.showMessage("Product deleted successfully", "info");
-        }
-    
+            $('#exampleModal').modal('show');
+        });
+  }
+
+  editProduct(e)
+  {
+        this.resetModal();
+        $('.modal-title').append('Edit product');
+        $('.modal-body').append(`<form id="product-form" class="card-body">
+                                    <div class="form-group">
+                                        <input type="text" id="mark" class="form-control" placeholder="Product Mark" />
+                                    </div>
+                                    <div class="form-group">
+                                        <input type="text" id="name" class="form-control" placeholder="Product Name" />
+                                    </div>
+                                    <div class="form-group">
+                                        <input type="number" step="0.01" id="price" min="0" class="form-control" placeholder="Product Price" />
+                                    </div>
+                                </form>`);
+        $('.saveChanges').click(()=>{
+            $('#exampleModal').modal('hide');
+        });
   }
 
   showMessage(message, css)
   {
-        $(".container").prepend(`<div class="alert alert-${css} mt-2">${message}</div>`);
+        $(".container").prepend(`<div class="alert alert-${css} mt-2" role="alert">${message}</div>`);
         setTimeout(function(){
             $(".alert").slideUp(1000, ()=>{
                 $(".alert").remove();
             })
-        }, 3000);
+        }, 2500);
   }
 }
 
@@ -49,13 +92,14 @@ class UI {
 $("#product-form").submit(function(e) {
     e.preventDefault();
 
+    const mark = $("#mark").val();
     const name = $("#name").val();
     const price = $("#price").val();
 
-    const product = new Product(name, price);
+    const product = new Product(mark, name, price);
 
     const ui = new UI();
-    if(name === '' || price === "")
+    if(mark === '' || name === '' || price === "")
     {
         return ui.showMessage("Complete fields please", "danger");
     }
@@ -68,12 +112,22 @@ $("#product-form").submit(function(e) {
 
 /*
     Delete product
+    Esta mal el evento click
 */
-
 $('#product-list').click((e)=>{
     e.preventDefault();
 
     const ui = new UI();
-    ui.deleteProduct(e.target);
+    
+    if(e.target.name === "delete")
+    {
+        console.log($(e).parent().parent().attr('id'));
+        ui.deleteProduct(e.target);
+    }
+    if(e.target.name === "edit")
+    {
+        console.log($(e).parent().parent().attr('id'));
+        ui.editProduct(e.target);
+    }
 
 });
